+++ eras/dijkstra/impl/CHANGELOG.md
@@ -95,6 +95,9 @@
 
 ### `testlib`
 
+* Add `InjectRuleFailure "LEDGER" DijkstraSubUtxowPredFailure era` as a superclass of `DijkstraEraImp`
+* Add `mkTopTxWithSubTxs`, `traverseSubTxs` and `withPostFixupSubTxs`
+* Add `Test.Cardano.Ledger.Dijkstra.Imp.SubUtxowSpec`
 * Add `switchTxToLegacyMode` helper
 * Add `balanceSubTransactions`
 * Expose `fixupSubTransactions`